requirements.txt
requests>=2.31.0
playwright>=1.43.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
